Fixed unit/integration testing; fixed a few linting errors;
