+++ example/file_input.html
@@ -101,7 +101,7 @@
         &copy; Copyright by Christoph Oberhofer
     </p>
 </footer>
-<script src="../src/vendor/jquery-1.9.0.min.js" type="text/javascript"></script>
+<script src="vendor/jquery-1.9.0.min.js" type="text/javascript"></script>
 <script src="../dist/quagga.js" type="text/javascript"></script>
 <script src="file_input.js" type="text/javascript"></script>
 </body>
