+++ karma-integration.conf.js
@@ -1,32 +1,43 @@
 module.exports = function(config) {
-  config.set({
-    basePath: '',
-    frameworks: ['mocha', 'requirejs', 'chai', 'sinon', 'sinon-chai'],
-    files: [
-      'test-main.js',
-      'src/typedefs.js',
-      {pattern: 'node_modules/async/lib/async.js', included: false},
-      {pattern: 'node_modules/gl-matrix/dist/gl-matrix-min.js', included: false},
-      {pattern: 'src/*.js', included: false},
-      {pattern: 'spec/**/*integration.spec.js', included: false},
-      {pattern: 'test/**/*.*', included: false}
-    ],
-    exclude: [
-    ],
-    plugins: [
-        'karma-chrome-launcher',
-        'karma-mocha',
-        'karma-requirejs',
-        'karma-chai',
-        'karma-sinon',
-        'karma-sinon-chai'
-    ],
-    reporters: ['progress'],
-    port: 9876,
-    colors: true,
-    logLevel: config.LOG_INFO,
-    autoWatch: true,
-    browsers: ['Chrome'],
-    singleRun: false
-  });
+    config.set({
+        basePath: '',
+        frameworks: ['mocha', 'chai', 'sinon', 'sinon-chai'],
+        files: [
+            'test/test-main-integration.js',
+            {pattern: 'test/integration/**/*.js', included: false},
+            {pattern: 'test/fixtures/**/*.*', included: false}
+        ],
+        preprocessors: {
+            'test/test-main-integration.js': ['webpack']
+        },
+        webpack: {
+            module: {
+                preLoaders: [
+                    {
+                        test: /\.js$/,
+                        exclude: [
+                            /node_modules/
+                        ],
+                        loader: 'babel'
+                    }
+                ]
+            }
+        },
+        plugins: [
+            'karma-chrome-launcher',
+            'karma-mocha',
+            'karma-requirejs',
+            'karma-chai',
+            'karma-sinon',
+            'karma-sinon-chai',
+            require('karma-webpack')
+        ],
+        reporters: ['progress'],
+        port: 9876,
+        colors: true,
+        logLevel: config.LOG_INFO,
+        autoWatch: true,
+        browsers: ['Chrome'],
+        singleRun: false
+    });
 };
