+++ package.json
@@ -10,12 +10,14 @@
     "babel-eslint": "^4.1.3",
     "babel-loader": "^5.3.2",
     "chai": "^3.2.0",
+    "core-js": "^1.2.1",
     "grunt": "^0.4.5",
     "grunt-contrib-jshint": "^0.11.3",
     "grunt-contrib-nodeunit": "^0.4.1",
     "grunt-contrib-uglify": "^0.9.2",
     "grunt-karma": "^0.12.1",
     "grunt-requirejs": "^0.4.2",
+    "isparta-loader": "^1.0.0",
     "karma": "^0.13.9",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "^0.2.0",
@@ -25,6 +27,8 @@
     "karma-requirejs": "~0.2.2",
     "karma-sinon": "^1.0.4",
     "karma-sinon-chai": "~0.2.0",
+    "karma-source-map-support": "^1.1.0",
+    "karma-webpack": "^1.7.0",
     "mocha": "^2.3.2",
     "sinon": "^1.16.1",
     "webpack": "^1.12.2",
